Split release candidate modules
